Repository Status Updates, Initial Submissions to Draft/Interim

+++ scripts/find_orphans.py
@@ -17,13 +17,20 @@ def main():
         upstream_ids = elements_index.find_upstream_ids(oval_ids, set())
         downstream_ids = elements_index.find_downstream_ids(oval_ids, set())
 
-        message("INFO", "OVAL ID '%s' had %i upstream ids and %i downstream ids" % (oval_id, len(upstream_ids), len(downstream_ids)))
+        #message("INFO", "OVAL ID '%s' had %i upstream ids and %i downstream ids" % (oval_id, len(upstream_ids), len(downstream_ids)))
 
         if len(upstream_ids) == 0 and len(downstream_ids) == 0:
+            message("INFO", "Found Orphan '%s'" % oval_id)
             orphans[oval_id] = elempath
 
+    response = input("\n :::: Remove all orphans? (N[o] / y[es]): ")
+    if response != 'y' and response != 'Y':
+        return
+
+    message("INFO", "\n\n")
+    
     for ok in orphans.keys():
-        message("INFO", "Found Orphan '%s'" % ok)
+        message("INFO", "Deleting Orphan '%s'" % ok)
         os.remove(orphans[ok])
 
 def message(label, message, silent=False):
